feat: add ProfilePage and LoginPage components, update routing and configuration
