+++ client/index.html
@@ -6,8 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Instaflop</title>
   </head>
-  <body>
-    <div id="app"></div>
+  <body id="app">
     <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
