+++ client/src/router.ts
@@ -1,41 +1,41 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import useAuthStore from './stores/auth-store';
+import useAuthStore from '@/stores/auth-store';
 
 const router = createRouter({
     history: createWebHistory(),
     routes: [
         {
             path: '/',
-            component: () => import('./pages/LandingPage.vue')
+            component: () => import('@/pages/LandingPage.vue')
         },
         {
 
             path: '/home',
-            component: () => import('./pages/HomePage.vue')
+            component: () => import('@/pages/home/HomePage.vue')
         },
         {
             path: '/login',
-            component: () => import('./pages/LoginPage.vue')
+            component: () => import('@/pages/login/LoginPage.vue')
         },
         {
             path: '/register',
-            component: () => import('./pages/RegisterPage.vue')
+            component: () => import('@/pages/register/RegisterPage.vue')
         },
         {
-            path: '/dashboard',
-            component: () => import('./pages/ProfilePage.vue'),
+            path: '/profile/:id',
+            component: () => import('@/pages/profile/ProfilePage.vue'),
             meta: { requiresAuth: true },
         },
         {
             path: '/:pathMatch(.*)',
-            component: () => import('./pages/NotFoundPage.vue')
+            component: () => import('@/pages/NotFoundPage.vue')
         }
     ],
 });
 
 router.beforeEach((to, _, next) => {
     const authStore = useAuthStore();
-    const isLoggedIn = authStore.getToken ?? false;
+    const isLoggedIn = authStore.getUser?.username ?? false;
 
     if (to.meta.requiresAuth && !isLoggedIn) {
         return next('/login');
